test(cli): run ordinary run-command semantics in process
Part of #2387. The maka run test suite paid a Node subprocess startup per
assertion, including for scenarios that only exercise argument handling,
session selection, and outcome-to-exit-code mapping through the injectable
MakaRunDeps seam.

- Extract the scenario fake from run-command-fixture.ts into
  run-command-fake.ts, parameterized by an options object instead of
  environment variables; the fixture becomes a thin subprocess wrapper
  over the same fake, so the two routes cannot drift apart.
- Move 21 ordinary-semantics tests in process through runMakaTextCli with
  injected stdin/stdout/stderr; assertions are unchanged.
- Retain real-subprocess coverage for every distinct process contract:
  piped non-TTY stdin (3), SIGINT delivery with exit 130 (2), and the
  fail-closed sandbox boundary path plus exit codes and stdout observed
  through a real process boundary (1).
- The graph-runtime-error scenario's "graph wait must not run" marker now
  travels through the thrown error message so the negative assertion stays
  observable on the captured stderr channel.

Timing (node --test dist/__tests__/run-command.test.js, local):
before 15.06s, after 3.60s. 34/34 tests pass, 5 consecutive rounds.
